feat: migrate checkCSP function to background
It is not executed in offscreen, because it is only meaningful to execute in the background.
