+++ src/workflowEngine/helper.js
@@ -1,6 +1,7 @@
-import browser from 'webextension-polyfill';
-import { customAlphabet } from 'nanoid/non-secure';
 import BrowserAPIService from '@/service/browser-api/BrowserAPIService';
+import { MessageListener } from '@/utils/message';
+import { customAlphabet } from 'nanoid/non-secure';
+import browser from 'webextension-polyfill';
 
 export function escapeElementPolicy(script) {
   if (window?.trustedTypes?.createPolicy) {
@@ -233,86 +234,22 @@ export async function checkCSPAndInject(
   { target, debugMode, options = {}, injectOptions = {} },
   callback
 ) {
-  const [isBlockedByCSP] = await browser.scripting.executeScript({
-    target,
-    func: () => {
-      return new Promise((resolve) => {
-        const escapePolicy = (script) => {
-          if (window?.trustedTypes?.createPolicy) {
-            const escapeElPolicy = window.trustedTypes.createPolicy(
-              'forceInner',
-              {
-                createHTML: (to_escape) => to_escape,
-                createScript: (to_escape) => to_escape,
-              }
-            );
-
-            return escapeElPolicy.createScript(script);
-          }
-
-          return script;
-        };
-        const eventListener = ({ srcElement }) => {
-          if (!srcElement || srcElement.id !== 'automa-csp') return;
-          srcElement.remove();
-          resolve(true);
-        };
-        document.addEventListener('securitypolicyviolation', eventListener);
-        const script = document.createElement('script');
-        script.id = 'automa-csp';
-        script.innerText = escapePolicy('console.log("...")');
-
-        setTimeout(() => {
-          document.removeEventListener(
-            'securitypolicyviolation',
-            eventListener
-          );
-          script.remove();
-          resolve(false);
-        }, 500);
-
-        document.body.appendChild(script);
-      });
-    },
-    world: 'MAIN',
-    ...(injectOptions || {}),
-  });
-
-  if (isBlockedByCSP.result) {
-    await new Promise((resolve) => {
-      chrome.debugger.attach({ tabId: target.tabId }, '1.3', resolve);
-    });
-
-    const jsCode = await callback();
-    const execResult = await sendDebugCommand(
-      target.tabId,
-      'Runtime.evaluate',
+  try {
+    const result = await MessageListener.sendMessage(
+      'check-csp-and-inject',
       {
-        expression: jsCode,
-        userGesture: true,
-        awaitPromise: true,
-        returnByValue: true,
-        ...(options || {}),
-      }
+        target,
+        debugMode,
+        callback: callback.toString(),
+        options,
+        injectOptions,
+      },
+      'background'
     );
-
-    if (!debugMode) await chrome.debugger.detach({ tabId: target.tabId });
-
-    if (!execResult || !execResult.result) {
-      throw new Error('Unable execute code');
-    }
-
-    if (execResult.result.subtype === 'error') {
-      throw new Error(execResult.result.description);
-    }
-
-    return {
-      isBlocked: true,
-      value: execResult.result.value || null,
-    };
+    return result;
+  } catch (err) {
+    return { isBlocked: true, value: null };
   }
-
-  return { isBlocked: false, value: null };
 }
 
 function fallbackCopyTextToClipboard(text) {
@@ -336,6 +273,7 @@ function fallbackCopyTextToClipboard(text) {
 
   document.body.removeChild(textArea);
 }
+
 export function copyTextToClipboard(text) {
   return new Promise((resolve, reject) => {
     if (!navigator.clipboard) {
